refactor(zoe): separate zcf into zygotable methods
